Refactor/split schema and add flow types
