--- todo/data/schema.js
@@ -1,297 +1,0 @@
-/**
- * This file provided by Facebook is for non-commercial testing and evaluation
- * purposes only.  Facebook reserves all rights not expressly granted.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
- * FACEBOOK BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
- * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
- * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-import {
-  GraphQLBoolean,
-  GraphQLID,
-  GraphQLInt,
-  GraphQLList,
-  GraphQLNonNull,
-  GraphQLObjectType,
-  GraphQLSchema,
-  GraphQLString,
-} from 'graphql';
-
-import {
-  connectionArgs,
-  connectionDefinitions,
-  connectionFromArray,
-  cursorForObjectInConnection,
-  fromGlobalId,
-  globalIdField,
-  mutationWithClientMutationId,
-  nodeDefinitions,
-  toGlobalId,
-} from 'graphql-relay';
-
-import {
-  Todo,
-  User,
-  USER_ID,
-  addTodo,
-  changeTodoStatus,
-  getTodo,
-  getTodos,
-  getUser,
-  markAllTodos,
-  removeCompletedTodos,
-  removeTodo,
-  renameTodo,
-} from './database';
-
-const {nodeInterface, nodeField} = nodeDefinitions(
-  globalId => {
-    const {type, id} = fromGlobalId(globalId);
-    if (type === 'Todo') {
-      return getTodo(id);
-    } else if (type === 'User') {
-      return getUser(id);
-    }
-    return null;
-  },
-  obj => {
-    if (obj instanceof Todo) {
-      return GraphQLTodo;
-    } else if (obj instanceof User) {
-      return GraphQLUser;
-    }
-    return null;
-  },
-);
-
-const GraphQLTodo = new GraphQLObjectType({
-  name: 'Todo',
-  fields: {
-    id: globalIdField('Todo'),
-    text: {
-      type: GraphQLString,
-      resolve: obj => obj.text,
-    },
-    complete: {
-      type: GraphQLBoolean,
-      resolve: obj => obj.complete,
-    },
-  },
-  interfaces: [nodeInterface],
-});
-
-const {
-  connectionType: TodosConnection,
-  edgeType: GraphQLTodoEdge,
-} = connectionDefinitions({
-  name: 'Todo',
-  nodeType: GraphQLTodo,
-});
-
-const GraphQLUser = new GraphQLObjectType({
-  name: 'User',
-  fields: {
-    id: globalIdField('User'),
-    userId: {
-      type: GraphQLString,
-      resolve: () => USER_ID,
-    },
-    todos: {
-      type: TodosConnection,
-      args: {
-        status: {
-          type: GraphQLString,
-          defaultValue: 'any',
-        },
-        ...connectionArgs,
-      },
-      resolve: (obj, {status, ...args}) =>
-        connectionFromArray(getTodos(status), args),
-    },
-    totalCount: {
-      type: GraphQLInt,
-      resolve: () => getTodos().length,
-    },
-    completedCount: {
-      type: GraphQLInt,
-      resolve: () => getTodos('completed').length,
-    },
-  },
-  interfaces: [nodeInterface],
-});
-
-const Query = new GraphQLObjectType({
-  name: 'Query',
-  fields: {
-    user: {
-      type: GraphQLUser,
-      args: {
-        id: {type: GraphQLString},
-      },
-      resolve: (root, {id}) => getUser(id),
-    },
-    node: nodeField,
-  },
-});
-
-const GraphQLAddTodoMutation = mutationWithClientMutationId({
-  name: 'AddTodo',
-  inputFields: {
-    text: {type: new GraphQLNonNull(GraphQLString)},
-    userId: {type: new GraphQLNonNull(GraphQLID)},
-  },
-  outputFields: {
-    todoEdge: {
-      type: GraphQLTodoEdge,
-      resolve: ({localTodoId}) => {
-        const todo = getTodo(localTodoId);
-        return {
-          cursor: cursorForObjectInConnection(getTodos(), todo),
-          node: todo,
-        };
-      },
-    },
-    user: {
-      type: GraphQLUser,
-      resolve: ({userId}) => getUser(userId),
-    },
-  },
-  mutateAndGetPayload: ({text, userId}) => {
-    const localTodoId = addTodo(text);
-    return {localTodoId, userId};
-  },
-});
-
-const GraphQLChangeTodoStatusMutation = mutationWithClientMutationId({
-  name: 'ChangeTodoStatus',
-  inputFields: {
-    complete: {type: new GraphQLNonNull(GraphQLBoolean)},
-    id: {type: new GraphQLNonNull(GraphQLID)},
-    userId: {type: new GraphQLNonNull(GraphQLID)},
-  },
-  outputFields: {
-    todo: {
-      type: GraphQLTodo,
-      resolve: ({localTodoId}) => getTodo(localTodoId),
-    },
-    user: {
-      type: GraphQLUser,
-      resolve: ({userId}) => getUser(userId),
-    },
-  },
-  mutateAndGetPayload: ({id, complete, userId}) => {
-    const localTodoId = fromGlobalId(id).id;
-    changeTodoStatus(localTodoId, complete);
-    return {localTodoId, userId};
-  },
-});
-
-const GraphQLMarkAllTodosMutation = mutationWithClientMutationId({
-  name: 'MarkAllTodos',
-  inputFields: {
-    complete: {type: new GraphQLNonNull(GraphQLBoolean)},
-    userId: {type: new GraphQLNonNull(GraphQLID)},
-  },
-  outputFields: {
-    changedTodos: {
-      type: new GraphQLList(GraphQLTodo),
-      resolve: ({changedTodoLocalIds}) => changedTodoLocalIds.map(getTodo),
-    },
-    user: {
-      type: GraphQLUser,
-      resolve: ({userId}) => getUser(userId),
-    },
-  },
-  mutateAndGetPayload: ({complete, userId}) => {
-    const changedTodoLocalIds = markAllTodos(complete);
-    return {changedTodoLocalIds, userId};
-  },
-});
-
-// TODO: Support plural deletes
-const GraphQLRemoveCompletedTodosMutation = mutationWithClientMutationId({
-  name: 'RemoveCompletedTodos',
-  inputFields: {
-    userId: {type: new GraphQLNonNull(GraphQLID)},
-  },
-  outputFields: {
-    deletedTodoIds: {
-      type: new GraphQLList(GraphQLString),
-      resolve: ({deletedTodoIds}) => deletedTodoIds,
-    },
-    user: {
-      type: GraphQLUser,
-      resolve: ({userId}) => getUser(userId),
-    },
-  },
-  mutateAndGetPayload: ({userId}) => {
-    const deletedTodoLocalIds = removeCompletedTodos();
-    const deletedTodoIds = deletedTodoLocalIds.map(
-      toGlobalId.bind(null, 'Todo'),
-    );
-    return {deletedTodoIds, userId};
-  },
-});
-
-const GraphQLRemoveTodoMutation = mutationWithClientMutationId({
-  name: 'RemoveTodo',
-  inputFields: {
-    id: {type: new GraphQLNonNull(GraphQLID)},
-    userId: {type: new GraphQLNonNull(GraphQLID)},
-  },
-  outputFields: {
-    deletedTodoId: {
-      type: GraphQLID,
-      resolve: ({id}) => id,
-    },
-    user: {
-      type: GraphQLUser,
-      resolve: ({userId}) => getUser(userId),
-    },
-  },
-  mutateAndGetPayload: ({id, userId}) => {
-    const localTodoId = fromGlobalId(id).id;
-    removeTodo(localTodoId);
-    return {id, userId};
-  },
-});
-
-const GraphQLRenameTodoMutation = mutationWithClientMutationId({
-  name: 'RenameTodo',
-  inputFields: {
-    id: {type: new GraphQLNonNull(GraphQLID)},
-    text: {type: new GraphQLNonNull(GraphQLString)},
-  },
-  outputFields: {
-    todo: {
-      type: GraphQLTodo,
-      resolve: ({localTodoId}) => getTodo(localTodoId),
-    },
-  },
-  mutateAndGetPayload: ({id, text}) => {
-    const localTodoId = fromGlobalId(id).id;
-    renameTodo(localTodoId, text);
-    return {localTodoId};
-  },
-});
-
-const Mutation = new GraphQLObjectType({
-  name: 'Mutation',
-  fields: {
-    addTodo: GraphQLAddTodoMutation,
-    changeTodoStatus: GraphQLChangeTodoStatusMutation,
-    markAllTodos: GraphQLMarkAllTodosMutation,
-    removeCompletedTodos: GraphQLRemoveCompletedTodosMutation,
-    removeTodo: GraphQLRemoveTodoMutation,
-    renameTodo: GraphQLRenameTodoMutation,
-  },
-});
-
-export const schema = new GraphQLSchema({
-  query: Query,
-  mutation: Mutation,
-});
