+++ todo/data/schema/mutations/ChangeTodoStatusMutation.js
@@ -1,0 +1,57 @@
+// @flow
+/* graphql-relay doesn't export types, and isn't in flow-typed.  This gets too messy */
+/* eslint flowtype/require-return-type: 'off' */
+/**
+ * This file provided by Facebook is for non-commercial testing and evaluation
+ * purposes only.  Facebook reserves all rights not expressly granted.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+ * FACEBOOK BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+ * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+import {fromGlobalId, mutationWithClientMutationId} from 'graphql-relay';
+import {GraphQLBoolean, GraphQLID, GraphQLNonNull} from 'graphql';
+import {GraphQLTodo, GraphQLUser} from '../nodes';
+import {changeTodoStatus, getTodo, getUser, Todo, User} from '../../database';
+
+type Input = {|
+  +complete: boolean,
+  +id: string,
+  +userId: string,
+|};
+
+type Payload = {|
+  +todoId: string,
+  +userId: string,
+|};
+
+const ChangeTodoStatusMutation = mutationWithClientMutationId({
+  name: 'ChangeTodoStatus',
+  inputFields: {
+    complete: {type: new GraphQLNonNull(GraphQLBoolean)},
+    id: {type: new GraphQLNonNull(GraphQLID)},
+    userId: {type: new GraphQLNonNull(GraphQLID)},
+  },
+  outputFields: {
+    todo: {
+      type: GraphQLTodo,
+      resolve: ({todoId}: Payload): ?Todo => getTodo(todoId),
+    },
+    user: {
+      type: GraphQLUser,
+      resolve: ({userId}: Payload): ?User => getUser(userId),
+    },
+  },
+  mutateAndGetPayload: ({id, complete, userId}: Input): Payload => {
+    const todoId = fromGlobalId(id).id;
+    changeTodoStatus(todoId, complete);
+
+    return {todoId, userId};
+  },
+});
+
+export {ChangeTodoStatusMutation};
