+++ todo/data/schema/mutations/RemoveCompletedTodosMutation.js
@@ -1,0 +1,57 @@
+// @flow
+/* graphql-relay doesn't export types, and isn't in flow-typed.  This gets too messy */
+/* eslint flowtype/require-return-type: 'off' */
+/**
+ * This file provided by Facebook is for non-commercial testing and evaluation
+ * purposes only.  Facebook reserves all rights not expressly granted.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+ * FACEBOOK BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+ * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+import {mutationWithClientMutationId, toGlobalId} from 'graphql-relay';
+import {GraphQLID, GraphQLList, GraphQLNonNull, GraphQLString} from 'graphql';
+import {GraphQLUser} from '../nodes';
+import {getUser, removeCompletedTodos, User} from '../../database';
+
+type Input = {|
+  +userId: string,
+|};
+
+type Payload = {|
+  +deletedTodoIds: $ReadOnlyArray<string>,
+  +userId: string,
+|};
+
+const RemoveCompletedTodosMutation = mutationWithClientMutationId({
+  name: 'RemoveCompletedTodos',
+  inputFields: {
+    userId: {type: new GraphQLNonNull(GraphQLID)},
+  },
+  outputFields: {
+    deletedTodoIds: {
+      type: new GraphQLList(GraphQLString),
+      resolve: ({deletedTodoIds}: Payload): $ReadOnlyArray<string> =>
+        deletedTodoIds,
+    },
+    user: {
+      type: GraphQLUser,
+      resolve: ({userId}: Payload): ?User => getUser(userId),
+    },
+  },
+  mutateAndGetPayload: ({userId}: Input): Payload => {
+    const deletedTodoLocalIds = removeCompletedTodos();
+
+    const deletedTodoIds = deletedTodoLocalIds.map(
+      toGlobalId.bind(null, 'Todo'),
+    );
+
+    return {deletedTodoIds, userId};
+  },
+});
+
+export {RemoveCompletedTodosMutation};
