+++ todo/data/schema/mutations/RemoveTodoMutation.js
@@ -1,0 +1,55 @@
+// @flow
+/* graphql-relay doesn't export types, and isn't in flow-typed.  This gets too messy */
+/* eslint flowtype/require-return-type: 'off' */
+/**
+ * This file provided by Facebook is for non-commercial testing and evaluation
+ * purposes only.  Facebook reserves all rights not expressly granted.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+ * FACEBOOK BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+ * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+import {mutationWithClientMutationId, fromGlobalId} from 'graphql-relay';
+import {GraphQLID, GraphQLNonNull} from 'graphql';
+import {GraphQLUser} from '../nodes';
+import {getUser, removeTodo, User} from '../../database';
+
+type Input = {|
+  +id: string,
+  +userId: string,
+|};
+
+type Payload = {|
+  +id: string,
+  +userId: string,
+|};
+
+const RemoveTodoMutation = mutationWithClientMutationId({
+  name: 'RemoveTodo',
+  inputFields: {
+    id: {type: new GraphQLNonNull(GraphQLID)},
+    userId: {type: new GraphQLNonNull(GraphQLID)},
+  },
+  outputFields: {
+    deletedTodoId: {
+      type: GraphQLID,
+      resolve: ({id}: Payload): string => id,
+    },
+    user: {
+      type: GraphQLUser,
+      resolve: ({userId}: Payload): ?User => getUser(userId),
+    },
+  },
+  mutateAndGetPayload: ({id, userId}: Input): Payload => {
+    const localTodoId = fromGlobalId(id).id;
+    removeTodo(localTodoId);
+
+    return {id, userId};
+  },
+});
+
+export {RemoveTodoMutation};
