+++ todo/data/schema/queries/UserQuery.js
@@ -1,0 +1,32 @@
+// @flow
+/* graphql-relay doesn't export types, and isn't in flow-typed.  This gets too messy */
+/* eslint flowtype/require-return-type: 'off' */
+/**
+ * This file provided by Facebook is for non-commercial testing and evaluation
+ * purposes only.  Facebook reserves all rights not expressly granted.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+ * FACEBOOK BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+ * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+import {GraphQLString} from 'graphql';
+import {GraphQLUser} from '../nodes';
+import {User, getUser} from '../../database';
+
+type Input = {
+  +id: string,
+};
+
+const UserQuery = {
+  type: GraphQLUser,
+  args: {
+    id: {type: GraphQLString},
+  },
+  resolve: (root: {}, {id}: Input): ?User => getUser(id),
+};
+
+export {UserQuery};
